+++ CHANGELOG.md
@@ -23,8 +23,9 @@
 - Added Scope and Breadcrumb ring buffer #109
 - Added Hub to SDK #113
 - Ref: Hub passes the Scope to SentryClient
-- Feat: sentry options #116
+- Feature: sentry options #116
 - Ref: SentryId generates UUID
+- Ref: Event now is SentryEvent and added GPU
 
 # `package:sentry` changelog